refactor: init PlayerPlugin from main.rs
- created a new module
- refracted code from main.rs into player.rs
- initialized plugin there to be used in main.rs
- removed .add_startup_system_to_stage() as it's already in the plugin
